[TECH] 0.2.0 Rework URL to MMSDMFile and NEMDEFile (#11)
* cleanup - remove unused files, move in development current to poc

* remove notebooks

* version bump, remove old package stuff

* mmsdm refactor

* nemde neu working

* nemde

* rm nemde neu

* add static typing to CI

* fix circular import

* fix nemde download

* fix static

* add dry run to cli

* finish changelog

+++ nemdata/mmsdm.py
@@ -1,79 +1,149 @@
-from pathlib import Path
+import datetime
+import pathlib
+import typing
+import warnings
 
+import numpy as np
 import pandas as pd
+import pydantic
+import requests
 from rich import print
 
-from nemdata.utils import URL, add_interval_cols, download_zipfile_from_url, unzip
-
-reports = {
-    "predispatch": {
-        "report": "PREDISPATCHPRICE",
-        "directory": "PREDISP_ALL_DATA",
-        "dt-cols": ["LASTCHANGED", "DATETIME"],
-        "interval-col": "DATETIME",
-        "freq": "30T",
-    },
-    "unit-scada": {
-        "report": "DISPATCH_UNIT_SCADA",
-        "directory": "DATA",
-        "interval-col": "SETTLEMENTDATE",
-        "dt-cols": ["SETTLEMENTDATE"],
-        "freq": "5T",
-    },
-    "trading-price": {
-        "report": "TRADINGPRICE",
-        "directory": "DATA",
-        "dt-cols": ["SETTLEMENTDATE"],
-        "interval-col": "SETTLEMENTDATE",
-        "freq": "30T",
-    },
-}
+from nemdata import utils
+from nemdata.config import DEFAULT_BASE_DIRECTORY
 
 
-def make_report_url(year, month, report, directory, report_id, base_dir):
-    #  zero pad the month
-    month = str(month).zfill(2)
-    prefix = f"https://www.nemweb.com.au/Data_Archive/Wholesale_Electricity/MMSDM/{year}/MMSDM_{year}_{month}/MMSDM_Historical_Data_SQLLoader"
+class VariableFrequency(pydantic.BaseModel):
+    frequency_minutes_before: int
+    frequency_minutes_after: int
+    transition_datetime: datetime.datetime
 
-    home = base_dir / report_id / f"{year}-{month}"
-    home.mkdir(exist_ok=True, parents=True)
 
-    return URL(
-        url=f"{prefix}/{directory}/PUBLIC_DVD_{report}_{year}{month}010000.zip",
-        year=year,
-        month=month,
-        report=report,
-        csv=f"PUBLIC_DVD_{report}_{year}{month}010000.CSV",
-        xml=None,
-        home=home,
+class MMSDMTable(pydantic.BaseModel):
+    name: str
+    table: str
+    directory: str
+    datetime_columns: typing.Union[list[str], None] = None
+    interval_column: typing.Union[str, None] = None
+    frequency: typing.Union[int, VariableFrequency, None] = None
+
+
+mmsdm_tables = [
+    MMSDMTable(
+        name="predispatch",
+        table="PREDISPATCHPRICE",
+        directory="PREDISP_ALL_DATA",
+        datetime_columns=["LASTCHANGED", "DATETIME"],
+        interval_column="DATETIME",
+        frequency=30,
+    ),
+    MMSDMTable(
+        name="unit-scada",
+        table="DISPATCH_UNIT_SCADA",
+        directory="DATA",
+        datetime_columns=["LASTCHANGED", "SETTLEMENTDATE"],
+        interval_column="SETTLEMENTDATE",
+        frequency=5,
+    ),
+    MMSDMTable(
+        name="trading-price",
+        table="TRADINGPRICE",
+        directory="DATA",
+        datetime_columns=["SETTLEMENTDATE"],
+        interval_column="SETTLEMENTDATE",
+        frequency=VariableFrequency(
+            frequency_minutes_before=30,
+            transition_datetime="2021-10-01T00:05:00",
+            frequency_minutes_after=5,
+        ),
+    ),
+]
+
+
+class MMSDMFile(pydantic.BaseModel):
+    year: int
+    month: int
+    table: MMSDMTable
+    url: str
+    csv_name: str
+    data_directory: pathlib.Path
+    zipfile_path: pathlib.Path
+
+
+def find_mmsdm_table(name: str) -> MMSDMTable:
+    """finds a MMSDMTable by it's `name`"""
+    for table in mmsdm_tables:
+        if table.name == name:
+            return table
+    raise ValueError(
+        f"MMSDMTable {name} not found - tables available are {[table.name for table in mmsdm_tables]}"
     )
 
 
-def make_many_report_urls(start, end, report_id, base_dir):
-    report = reports[report_id]
+def make_many_mmsdm_files(
+    start: str, end: str, table: MMSDMTable, base_directory: pathlib.Path
+) -> list[MMSDMFile]:
+    """creates many MMSDMFiles - one for each month"""
+    table = find_mmsdm_table(table.name)
     months = pd.date_range(start=start, end=end, freq="MS")
 
-    urls = []
+    files = []
     for year, month in zip(months.year, months.month):
-        urls.append(
-            make_report_url(
-                year, month, report["report"], report["directory"], report_id, base_dir
+        files.append(
+            make_one_mmsdm_file(
+                year=year, month=month, table=table, base_directory=base_directory
             )
         )
-    return urls
+    return files
 
 
-def load_unzipped_report(url, path, skiprows=1, tail=-1):
-    path = path.parent / url.csv
+def make_one_mmsdm_file(
+    year: int, month: int, table: MMSDMTable, base_directory: pathlib.Path
+) -> MMSDMFile:
+    """creates a single MMSDMFile object that represents one file on the AEMO MMSDM website"""
+    #  zero pad the month - 3 -> 03
+    padded_month = str(month).zfill(2)
+
+    #  url to the zipfile on MMSDM website
+    url_prefix = f"https://www.nemweb.com.au/Data_Archive/Wholesale_Electricity/MMSDM/{year}/MMSDM_{year}_{padded_month}/MMSDM_Historical_Data_SQLLoader"
+    url = f"{url_prefix}/{table.directory}/PUBLIC_DVD_{table.table}_{year}{padded_month}010000.zip"
+
+    #  name of the CSV that comes out of the zipfile
+    csv_name = f"PUBLIC_DVD_{table.table}_{year}{padded_month}010000.CSV"
+
+    #  data directory where we will download data to
+    data_directory = base_directory / table.name / f"{year}-{padded_month}"
+    data_directory.mkdir(exist_ok=True, parents=True)
+
+    return MMSDMFile(
+        year=year,
+        month=month,
+        table=table,
+        url=url,
+        csv_name=csv_name,
+        data_directory=data_directory,
+        zipfile_path=data_directory / "raw.zip",
+    )
+
+
+def load_unzipped_mmsdm_file(
+    mmsdm_file: MMSDMFile, skiprows: int = 1, tail: int = -1
+) -> pd.DataFrame:
+    """read the CSV from an unzipped MMSDMFile"""
+    path = mmsdm_file.data_directory / mmsdm_file.csv_name
     #  remove first row via skiprows
     data = pd.read_csv(path, skiprows=skiprows)
     #  remove last row via iloc
     return data.iloc[:tail, :]
 
 
-def make_dt_cols(data, dt_cols):
-    dt_cols += ["interval-start", "interval-end", "timestamp"]
-    for col in dt_cols:
+def make_datetime_columns(data: pd.DataFrame, table: MMSDMTable) -> pd.DataFrame:
+    """cast the `mmsdm_table.datetime_columns` to datetime objects"""
+    datetime_columns = table.datetime_columns
+    assert datetime_columns
+    datetime_columns += ["interval-start", "interval-end"]
+
+    for col in datetime_columns:
         try:
             data[col] = pd.to_datetime(data[col])
         except KeyError:
@@ -81,39 +151,42 @@ def make_dt_cols(data, dt_cols):
     return data
 
 
-def download_mmsdm(start, end, report_id, base_dir):
-    urls = make_many_report_urls(start, end, report_id, base_dir)
+def download_mmsdm(
+    start: str,
+    end: str,
+    table_name: str,
+    base_directory: pathlib.Path = DEFAULT_BASE_DIRECTORY,
+    dry_run: bool = False,
+) -> pd.DataFrame:
+    """main for downloading MMSDMFiles"""
+    table = find_mmsdm_table(table_name)
+    files = make_many_mmsdm_files(start, end, table, base_directory)
 
-    output = []
-    for url in urls:
-        clean_fi = url.home / "clean.parquet"
+    dataset = []
+    for mmsdm_file in files:
+        clean_fi = mmsdm_file.data_directory / "clean.parquet"
         if clean_fi.exists():
-            print(f" [blue]exists[/] {clean_fi.parts[-5:]}")
+            print(f" [blue]EXISTS[/] {' '.join(clean_fi.parts[-5:])}")
             data = pd.read_parquet(clean_fi)
         else:
-            print(f" [blue]missing[/] {clean_fi.parts[-5:]}")
-            zf = download_zipfile_from_url(url)
-            unzip(zf)
+            print(f" [blue]MISSING[/] {' '.join(clean_fi.parts[-5:])}")
+            print(
+                f" [green]DOWNLOADING[/] {' '.join(mmsdm_file.zipfile_path.parts[-5:])}"
+            )
+            utils.download_zipfile(mmsdm_file)
+            utils.unzip(mmsdm_file.zipfile_path)
+            data = load_unzipped_mmsdm_file(mmsdm_file)
+            assert table.datetime_columns
+            data = make_datetime_columns(data, table)
+            data = utils.add_interval_column(data, table)
 
-            data = load_unzipped_report(url, zf)
+            if not dry_run:
+                print(f" [green]SAVING [/] {clean_fi}")
+                data.to_csv(clean_fi.with_suffix(".csv"))
+                data.to_parquet(clean_fi.with_suffix(".parquet"))
+        dataset.append(data)
+    return pd.concat(dataset, axis=0)
 
-            #  unpacking the report dict - must be better way...
-            report = reports[report_id]
-            timestamp_col = report["interval-col"]
 
-            data = make_dt_cols(data, report["dt-cols"])
-
-            #  accounting for AEMO stamping intervals at the end
-            #  usually intervals are stamped at the start
-            data = add_interval_cols(data, timestamp_col, report["freq"])
-
-            #  could check by assert difference == freq
-            print(f" [green]saving [/] {url.home.parts[-4:]}")
-            data.to_csv(clean_fi.with_suffix(".csv"))
-            data.to_parquet(clean_fi.with_suffix(".parquet"))
-
-        print("\n")
-
-        output.append(data)
-
-    return pd.concat(output, axis=0)
+if __name__ == "__main__":
+    download_mmsdm(start="2021-10", end="2021-11", table_name="trading-price")
